% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lavaan_results.R
\name{tidy_lavaan}
\alias{tidy_lavaan}
\title{Create a tidy results dataframe from a fitted lavaan object}
\usage{
tidy_lavaan(x, stdest = FALSE)
}
\arguments{
\item{x}{fitted lavaan object}

\item{std.est}{Do you want standardized estimates? Defaults to FALSE}
}
\value{
dataframe
}
\description{
This function returns a formatted, tidy dataframe from a fitted lavaan object.
The object assumes at least one latent variable measured with reflective
indicators, although can be a simple CFA with no structural components.
There is also the abilty to include an indirect effect if testing mediation.
Note the model type codes used internally to specify the type of model object.
  Structural parameters: 1
  Measurement parameters: 2
  Indirect effect parameters: 3
  Covariance parameters: 4
  Model fit parameters: 5
}
\examples{
tidy_lavaan()
tidy_lavaan(x, std.est = TRUE)

}
