% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lavaan_fit.R
\name{tidy_lavaanFit}
\alias{tidy_lavaanFit}
\title{Create a tidy model fit dataframe from a fitted lavaan object}
\usage{
tidy_lavaanFit(x)
}
\arguments{
\item{x}{fitted lavaan object}
}
\value{
dataframe
}
\description{
This function returns a tidy dataframe containing the Chi-squared statistic,
degrees of freedom, and p-value from a fitted lavaan object. The dataframe
is preformatted to align with the tidy_lavaan output format to easily
combine the tables together for presentation in a paper.
}
\examples{
tidy_lavaanFit()

}
